refactor: split cli-parts of config module to cli/
